Resolves #4 - Error when displaying Richard Kelly as director
There was an error in the loop that selects items to be browsed. On line 52 of `items/browse.php` the `$film` variable was being set to `null` but never checked. Why this happened only on Mr. Kelly is a mystery. Added a check and continue if `$film` is null at the top of the loop. All seems well.

Also change permission on numerous files to match the deployed permissions.